fixed scrolling and developed seperate routes for project_manager data handling
