+++ templates/add_strategic_project_manager_modal.html
@@ -6,7 +6,7 @@
                 <h4 class="modal-title">Add New Data</h4>
             </div>
             <div class="modal-body">
-                <form action="{{ url_for('projects.insert_project_manager') }}" method="post">
+                <form action="{{ url_for('project_manager.insert_project_manager') }}" method="post">
                     <div class="form-group">
                         <label>Name <span style="color: red;">*</span></label>
                         <input type="text" class="form-control" name="name" required>
