+++ templates/add_users_data_modal.html
@@ -33,7 +33,11 @@
                             <option value="admin">Administrator</option>
                             <option value="admin_projects">Admin Projects</option>
                             <option value="admin_gis">Admin GIS</option>
-                            <option value="admin_strats">Admin Strategic Planning</option>
+                            <option value="admin_strategic_planning">Admin Strategic Planning</option>
+                            <option value="admin_water">Admin Water</option>
+                            <option value="admin_sanitation">Admin Sanitation</option>
+                            <option value="admin_electromechanical">Admin Electromechanical</option>
+                            <option value="admin_water_quality">Admin Water Quality</option>
                             <option value="user">User</option>
                         </select>
                     </div>
