+++ templates/add_water_projects_modal.html
@@ -42,8 +42,8 @@
                         <input type="text" class="form-control" name="contractor">
                     </div>
                     <div class="form-group">
-                        <label>Year</label>
-                        <input type="text" class="form-control" name="year">
+                        <label>Year <span style="color: red;">*</span></label>
+                        <input type="text" class="form-control" name="year" required>
                     </div>
                     <div class="form-group">
                         <label>Date Contract Signed</label>
