+++ templates/electromechanical_projects.html
@@ -131,8 +131,8 @@
             // "info": false, // Disable info display for inline editing
             // "scrollY": '50vh', // Set a fixed height for vertical scrolling
             // "scrollCollapse": true, // Enable scrolling
-            "scrollX": true // Enable horizontal scrolling
-            // "autoWidth": false // Disable automatic column width calculation
+            "scrollX": true, // Enable horizontal scrolling
+            "autoWidth": false // Disable automatic column width calculation
         });
     });
 </script>
