+++ templates/gis.html
@@ -69,7 +69,8 @@
     $(document).ready(function() {
         $('#gis_table').DataTable({
             "scrollX": true, // Enable horizontal scrolling
-            "fixedHeader": true // Keeps header fixed when scrolling horizontally
+            "fixedHeader": true, // Keeps header fixed when scrolling horizontally
+            "autoWidth": false // Disable automatic column width calculation
         });
     });
 </script>
