+++ templates/gis_data.html
@@ -186,13 +186,13 @@
   $(document).ready(function() {
       $('#gis_output_table, #gis_activity_table, #gis_resp_person_table, #gis_task_table').DataTable({
           // "paging": true, // Disable pagination for inline editing
-          "ordering": true, // Disable sorting for inline editing
-          "fixedHeader": true // Keeps header fixed when scrolling horizontally
+          "ordering": false, // Disable sorting for inline editing
+          "fixedHeader": true, // Keeps header fixed when scrolling horizontally
           // "info": false, // Disable info display for inline editing
           // "scrollY": '50vh', // Set a fixed height for vertical scrolling
           // "scrollCollapse": true, // Enable scrolling
-          // "scrollX": true, // Enable horizontal scrolling
-          // "autoWidth": false // Disable automatic column width calculation
+          "scrollX": true, // Enable horizontal scrolling
+          "autoWidth": false // Disable automatic column width calculation
       });
   });
 </script>
