+++ templates/projects_data.html
@@ -195,7 +195,7 @@
 <script src="https://cdn.datatables.net/2.0.8/js/dataTables.bootstrap5.js"></script>
 <script>
     $(document).ready(function() {
-        $('#projects_data_table').DataTable({
+        $('#projects_data_table, #projects_sections_table, #projects_project_manager_table').DataTable({
             "fixedHeader": true, // Keeps header fixed when scrolling horizontally
             // "paging": true, // Disable pagination for inline editing
             "ordering": true, // Disable sorting for inline editing
@@ -203,21 +203,7 @@
             // "scrollY": '50vh', // Set a fixed height for vertical scrolling
             // "scrollCollapse": true, // Enable scrolling
             "scrollX": true, // Enable horizontal scrolling
-            // "autoWidth": false // Disable automatic column width calculation
-        });
-    });
-</script>
-<script>
-    $(document).ready(function() {
-        $('#projects_project_manager_table').DataTable({
-            "fixedHeader": true, // Keeps header fixed when scrolling horizontally
-            // "paging": true, // Disable pagination for inline editing
-            "ordering": true, // Disable sorting for inline editing
-            // "info": false, // Disable info display for inline editing
-            // "scrollY": '50vh', // Set a fixed height for vertical scrolling
-            // "scrollCollapse": true, // Enable scrolling
-            // "scrollX": true, // Enable horizontal scrolling
-            // "autoWidth": false // Disable automatic column width calculation
+            "autoWidth": false // Disable automatic column width calculation
         });
     });
 </script>
