+++ templates/strategic_planning.html
@@ -103,7 +103,7 @@
             // "scrollY": '50vh', // Set a fixed height for vertical scrolling
             // "scrollCollapse": true, // Enable scrolling
             "scrollX": true, // Enable horizontal scrolling
-            // "autoWidth": false // Disable automatic column width calculation
+            "autoWidth": false // Disable automatic column width calculation
         });
     });
 </script>
