+++ templates/strategic_planning_data.html
@@ -20,6 +20,42 @@
     <div class="row">
         <div class="col md-12">
             <div class="jumbotron data_container p-3">
+                <h2 class="left-aligned-text">Strategic Planning Project Managers
+                    <button type="button" class="btn btn-success float-right" data-toggle="modal" data-target="#myStrategicProjectManagerModal">Add New Data</button>
+                </h2>
+                <div class="table-responsive">
+                    <table id="strats_project_manager_table" class="table table-bordered table-striped table-hover">
+                        <thead>
+                            <tr class="table-dark">
+                                <th>No.</th>
+                                <th>Project Manager</th>
+                                <th>Actions</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            {% for project_manager in project_managers %}
+                            <tr>
+                                <td>{{ loop.index }}</td>
+                                <td>{{ project_manager.name or '' }}</td>
+                                <td>
+                                    <a href="{{ url_for('project_manager.update_projects_project_manager', project_manager_id=project_manager.id) }}" class="btn btn-warning btn-xs update-btn" data-toggle="modal" data-target="#updateStrategicProjectManagerModal{{ project_manager.id }}">Edit</a>
+                                    <a href="{{ url_for('project_manager.delete_projects_project_manager', project_manager_id=project_manager.id) }}" class="btn btn-danger btn-xs" onclick="return confirm('Are you sure you want to delete this record?')">Delete</a>
+                                </td>
+                                {% include 'add_strategic_project_manager_modal.html' %}
+                                {% include 'update_strategic_project_manager_modal.html' %}
+                            </tr>
+                            {% endfor %}
+                        </tbody>
+                    </table>
+                </div>
+            </div>
+        </div>  
+    </div>
+</div>
+<div class="stats_data_container">
+    <div class="row">
+        <div class="col md-12">
+            <div class="jumbotron data_container p-3">
                 <h2 class="left-aligned-text">Strategic Planning Data
                     <button type="button" class="btn btn-success float-right" data-toggle="modal" data-target="#myStrategicModal">Add New Data</button>
                 </h2>
@@ -83,42 +119,6 @@
         </div>  
     </div>
 </div>
-<div class="stats_data_container">
-    <div class="row">
-        <div class="col md-12">
-            <div class="jumbotron data_container p-3">
-                <h2 class="left-aligned-text">Strategic Planning Project Managers
-                    <button type="button" class="btn btn-success float-right" data-toggle="modal" data-target="#myStrategicProjectManagerModal">Add New Data</button>
-                </h2>
-                <div class="table-responsive">
-                    <table id="strats_project_manager_table" class="table table-bordered table-striped table-hover">
-                        <thead>
-                            <tr class="table-dark">
-                                <th>No.</th>
-                                <th>Project Manager</th>
-                                <th>Actions</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {% for project_manager in project_managers %}
-                            <tr>
-                                <td>{{ loop.index }}</td>
-                                <td>{{ project_manager.name or '' }}</td>
-                                <td>
-                                    <a href="{{ url_for('projects.update_projects_project_manager', project_manager_id=project_manager.id) }}" class="btn btn-warning btn-xs update-btn" data-toggle="modal" data-target="#updateStrategicProjectManagerModal{{ project_manager.id }}">Edit</a>
-                                    <a href="{{ url_for('projects.delete_projects_project_manager', project_manager_id=project_manager.id) }}" class="btn btn-danger btn-xs" onclick="return confirm('Are you sure you want to delete this record?')">Delete</a>
-                                </td>
-                                {% include 'add_strategic_project_manager_modal.html' %}
-                                {% include 'update_strategic_project_manager_modal.html' %}
-                            </tr>
-                            {% endfor %}
-                        </tbody>
-                    </table>
-                </div>
-            </div>
-        </div>  
-    </div>
-</div>
 <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/js/bootstrap.bundle.min.js"></script>
 <script src="https://cdn.datatables.net/2.0.8/js/dataTables.js"></script>
@@ -126,14 +126,14 @@
 <script>
     $(document).ready(function() {
         $('#strats_data_table, #strats_project_manager_table').DataTable({
-            "fixedHeader": true // Keeps header fixed when scrolling horizontally
+            //"fixedHeader": true, // Keeps header fixed when scrolling horizontally
             // "paging": true, // Disable pagination for inline editing
             // "ordering": false, // Disable sorting for inline editing
             // "info": false, // Disable info display for inline editing
             // "scrollY": '50vh', // Set a fixed height for vertical scrolling
             // "scrollCollapse": true, // Enable scrolling
-            // "scrollX": true, // Enable horizontal scrolling
-            // "autoWidth": false // Disable automatic column width calculation
+            "scrollX": true, // Enable horizontal scrolling
+            "autoWidth": false // Disable automatic column width calculation
         });
     });
 </script>
