+++ templates/update_projects_modal.html
@@ -50,8 +50,8 @@
                 <input type="text" class="form-control" name="contractor" value="{{ project_data.contractor }}">
             </div>
             <div class="form-group">
-                <label>Year</label>
-                <input type="text" class="form-control" name="year" value="{{ project_data.year }}">
+                <label>Year <span style="color: red;">*</span></label>
+                <input type="text" class="form-control" name="year" value="{{ project_data.year }} required">
             </div>
             <div class="form-group">
                 <label>Date Contract Signed</label>
