+++ templates/update_strategic_modal.html
@@ -10,7 +10,7 @@
           <form action="{{ url_for('strategic.update_strategic_data', strategic_data_id=strategic_data.task_id) }}" method="post">
             <div class="form-group">
                 <label>Task <span style="color: red;">*</span></label>
-                <input type="text" class="form-control" name="task" value="{{ strategic_data.task }}" required>
+                <input type="text" class="form-control" name="task" value="{{ strategic_data.task or '' }}" required>
             </div>
             <div class="form-group">
                 <label>Description </label>
