+++ templates/update_users_data_modal.html
@@ -36,7 +36,11 @@
                         'admin': 'Administrator',
                         'admin_projects': 'Admin Projects',
                         'admin_gis': 'Admin GIS',
-                        'admin_struts': 'Admin Strategic Planning',
+                        'admin_strategic_planning': 'Admin Strategic Planning',
+                        'admin_water': 'Admin Water',
+                        'admin_sanitation': 'Admin Sanitation',
+                        'admin_electromechanical': 'Admin Electromechanical',
+                        'admin_water_quality': 'Admin Water Quality',
                         'user': 'User'
                     } %}
