+++ templates/water_projects_data.html
@@ -161,27 +161,14 @@
 <script src="https://cdn.datatables.net/2.0.8/js/dataTables.bootstrap5.js"></script>
 <script>
     $(document).ready(function() {
-        $('#water_projects_data_table').DataTable({
+        $('#water_projects_data_table, #water_projects_project_manager_table').DataTable({
             // "paging": true, // Disable pagination for inline editing
             "ordering": true, // Disable sorting for inline editing
             // "info": false, // Disable info display for inline editing
             // "scrollY": '50vh', // Set a fixed height for vertical scrolling
             // "scrollCollapse": true, // Enable scrolling
             "scrollX": true, // Enable horizontal scrolling
-            // "autoWidth": false // Disable automatic column width calculation
-        });
-    });
-</script>
-<script>
-    $(document).ready(function() {
-        $('#water_projects_project_manager_table').DataTable({
-            // "paging": true, // Disable pagination for inline editing
-            "ordering": true, // Disable sorting for inline editing
-            // "info": false, // Disable info display for inline editing
-            // "scrollY": '50vh', // Set a fixed height for vertical scrolling
-            // "scrollCollapse": true, // Enable scrolling
-            // "scrollX": true, // Enable horizontal scrolling
-            // "autoWidth": false // Disable automatic column width calculation
+            "autoWidth": false // Disable automatic column width calculation
         });
     });
 </script>
